Tests/AK: Merge list tests
